Refactor runner_module() into two pieces
One piece still suitable for usage with argparse's "type" and another
suitable for internal use.  For the latter, we don't want to be
importing from nextstrain.cli.argparse or catching ArgumentTypeErrors.
